backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
jinja2>=3.1.0
aiofiles>=23.0.0

# Database
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0

# AI — OpenAI
openai>=1.35.0

# Document parsing
pypdf>=4.0.0
python-docx>=1.1.0
httpx>=0.27.0

# Knowledge base
rank-bm25>=0.2.2
python-frontmatter>=1.1.0
python-slugify>=8.0.0

# Config
pydantic-settings>=2.0.0
pydantic>=2.0.0

# Optional: semantic reranking (CPU-only, no GPU needed)
# fastembed>=0.3.0

# Dev / testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
